airflow/requirements.txt
apache-airflow==2.8.0
kafka-python==2.0.2
hdfs==2.7.3
requests==2.31.0
